Fix canonicalize for Windows (via dunce)
